Move parser to correct module
