+++ Cargo.toml
@@ -12,3 +12,4 @@ license = "MIT"
 maplit = "*"
 unicode_categories = "0.1"
 unicode-segmentation = "0.1"
+nom = "*"
