+++ src/parse/errors.rs
@@ -1,11 +1,11 @@
 //! Error handling in parsers
 
-use lex::{Expression, ExpressionType, Token, TokenType};
+use lex::{Token, TokenType};
+use parse::expression::{Expression, ExpressionType};
 
 /// Result given from main and expression parsers
 pub type ParseResult = Result<Expression, ParseError>;
 
-pub type TokenResult<'a> = Result<Token, ParseError>;
 
 /// Error given from parsers
 #[derive(Debug, Clone, PartialEq)]
