+++ src/parse/mod.rs
@@ -1,1 +1,9 @@
-mod types;
+mod errors;
+pub mod expression;
+mod parser;
+mod precedence;
+pub mod symbol;
+
+pub use self::errors::{ParseError, ParseResult};
+pub use self::parser::Parser;
+pub use self::precedence::Precedence;
